pyproject.toml
[project]
name = "rtmdet-object-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.1.3",
    "onnx>=1.17.0",
    "onnxruntime>=1.21.1",
    "onnxscript>=0.2.5",
    "pyaml>=25.1.0",
    "pydantic>=2.11.4",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "maturin>=1.8.4",
    "patchelf>=0.17.2.2",
    "pytest>=8.3.3",
]

[tool.pytest.ini_options]
pythonpath = [
    ".", "src",
]
